==> step2_articles_pre_processing.ipynb <==
import pandas as pd
from cleantext import clean

from nltk.corpus import stopwords
from nltk.probability import FreqDist
import numpy as np

def limpar_texto(text):
    
    return clean.clean(text,
                    fix_unicode=True,              
                    to_ascii=True,                 
                    lower=True,                    
                    no_line_breaks=True,
                    no_urls=True,
                    no_emails=True,     
                    no_phone_numbers=True,
                    no_numbers=True,      
                    no_digits=True,       
                    no_currency_symbols=True, 
                    no_punct=True,          
                    replace_with_punct="",
                    replace_with_url="",
                    replace_with_email="",
                    replace_with_phone_number="",
                    replace_with_number="",
                    replace_with_digit="",
                    replace_with_currency_symbol="",
                    lang="pt-br"
                )
    
def plot_frequencia_tokens(tokens):
    fd_words = FreqDist(tokens)
    fd_words.plot(20)
    
def remover_stopwords(text, personal_stop=[]):
    
    portugues_stops = stopwords.words('portuguese')

    text = text.split(' ')
    text = [x for x in text if x not in portugues_stops]
    text = [x for x in text if x not in personal_stop]
    return ' '.join(text)

def get_most_frequent_words(words, total=100):
    fd_words = FreqDist(words)

    words_to_remove = []
    counter = 0
    for word, i in fd_words.most_common():
        words_to_remove.append(word)
        counter += 1
        if counter > total:
            break
    
    return words_to_remove
    
def contador_palavras(texts):
    total = 0
    count_texts = 0
    max = 0
    min = 0
    for text in texts:
        count_texts += 1
        total_palavras = len(text.split(' '))
        total += total_palavras
        if total_palavras > max:
            max = total_palavras
        
        if min > total_palavras or min == 0:
            min = total_palavras

    return total, total/count_texts, max, min

def listar_palavras_do_Series(series):
    words_to_remove = []
    for text in series:
        for word in text.split(' '):
            words_to_remove.append(word)

    return words_to_remove



df_artigos = pd.read_csv('articles.csv')
df_novos_artigos = pd.read_excel("new_scrapped_articles/fixed_articles.xlsx")


df_artigos.head()

df_artigos.date = df_artigos.date.astype(str)

df_novos_artigos[df_novos_artigos.date == 'None'].shape, df_novos_artigos.shape

df_novos_artigos.drop(inplace=True, index=df_novos_artigos[df_novos_artigos.date == 'None'].index)

df_novos_artigos.date = pd.to_datetime(df_novos_artigos.date)

df_novos_artigos.date[1], df_artigos.date[1]

df_artigos.shape

new_data = []

for i, item in df_novos_artigos.iterrows():

    new_data.append([
            item.title, 
            item.text,
            item.date.strftime('%Y-%m-%d'),
            str(item.section).lower(),
            np.nan,
            item.url,
        ])



df_novos_artigos_2 = pd.DataFrame(columns=[
            'title', 
            'text',
            'date',
            'category',
            'subcategory',
            'link']
            , data=new_data
            )

df_novos_artigos_2.head()

df_new = pd.concat([df_artigos, df_novos_artigos_2], axis=0)
df_new

df_new.date.max(), df_new.date.min()

df_new.shape

df_new.head()
df_new.category.unique()

df_artigos.shape, df_new.shape

df_new.drop(index=df_new[df_new.text.isna()].index, inplace=True)
df_artigos.drop(index=df_artigos[df_artigos.text.isna()].index, inplace=True)

df_artigos.shape, df_new.shape

filtro = (df_artigos.text.str.lower().str.contains("etrobr"))
df_artigos[filtro].shape

filtro = (df_new.text.str.lower().str.contains("etrobr"))
df_new[filtro].shape

df_new = df_new[filtro]

df_new = df_new.drop(df_new[df_new.date.isnull()].index)


df_new['date_str'] = df_new.date.convert_dtypes(convert_string=True)

df_new.shape

df_new.head()

df_new.text = [limpar_texto(txt) for txt in df_new.text.replace('\n', ' ').replace('\t', ' ')]

contador_palavras(df_new.text)

df_new.text = [remover_stopwords(txt) for txt in df_new.text]

df_new['date_str'] = pd.to_datetime(df_new.date, errors='coerce')

df_new = df_new.drop(df_new[df_new.date_str.isnull()].index)
df_new.shape

df_new = df_new.drop(df_new[df_new.text.isnull()].index)
df_new.shape

df_new.to_csv('articles_por_data_sem_merge.csv', sep='\t')

select_header = ['date', 'title', 'text', 'link']

df_dados_etiquetados = pd.read_excel('dados_etiquetados.xlsx', sheet_name='dados', index_col=False)
df_articles_por_data_etiquetados = pd.merge(df_new[select_header], df_dados_etiquetados[['title', 'sobre_petrobras_diretamente']], on=['title'], how='left')

df_articles_por_data_etiquetados.dropna().to_csv('articles_por_data_etiquetados.csv', sep='\t', encoding='utf-8')

df_articles_petro = df_new[df_new.text.str.contains('Petrobras')]
df_articles_petro_grouped = df_new.groupby('date')

df_articles_por_data = pd.concat([df_articles_petro_grouped.title.apply(lambda x: ' '.join(x)), df_articles_petro_grouped.text.apply(lambda x: ' '.join(x)), df_articles_petro_grouped.link.apply(lambda x: ' - '.join(x))], axis=1)


df_articles_por_data_etiquetados.to_csv('articles_per_date.csv', sep='\t', encoding='utf-8')
