==> step3_data_merger.ipynb <==
import pandas as pd
import numpy as np
from datetime import date

import datetime


pontos_de_corte = 0.01

def merge_df(df_articles: pd.DataFrame, df_indices: pd.DataFrame) -> pd.DataFrame:
    
    clean_texts = [] 
    dt_temp = df_indices.index.min()

    df_articles = df_articles.sort_values('date').copy()

    for index, row in df_articles.iterrows():

        if not np.isnan(df_indices.reindex([row.date]).COEF_ANTERIOR[0]):
            dt_temp = row.date

        dt = dt_temp

        indice = resolve_indice(df_indices, dt)

        row_data = [dt, row.title, row.text, indice, df_indices.reindex([dt]).COMPARATIVO[0], row.link, df_indices.reindex([dt]).COEF[0], df_indices.reindex([dt]).COEF_ANTERIOR[0]]
        if('sobre_petrobras_diretamente' in df_articles.keys()):
            row_data.append(row.sobre_petrobras_diretamente)

        clean_texts.append(row_data)
        
    header = ['date', 'title', 'text', 'indice', 'comparativo', 'link', 'coef', 'coef_anterior']

    if('sobre_petrobras_diretamente' in df_articles.keys()):
        header.append('sobre_petrobras_diretamente')

    return pd.DataFrame(clean_texts, columns=header)

def resolve_indice(df_indices, dt):
    if df_indices.reindex([dt]).COMPARATIVO[0] >= pontos_de_corte:
        return 1
    elif df_indices.reindex([dt]).COMPARATIVO[0] <= (pontos_de_corte * -1):
        return -1
    else:
        return 0

df_articles = pd.read_csv('articles_per_date.csv', sep='\t', header=0)
df_articles_sem_merge = pd.read_csv('articles_por_data_sem_merge.csv', sep='\t', header=0)
df_articles_etiquetados = pd.read_csv('articles_por_data_etiquetados.csv', sep='\t', header=0)

df_articles_sem_merge.head()

Convert para data

df_datas = pd.to_datetime(df_articles.date)
df_articles.date = df_datas

df_datas = pd.to_datetime(df_articles_sem_merge.date)
df_articles_sem_merge.date = df_datas

df_articles_sem_merge.date.info()

Importa índices

df_indices = pd.read_excel('indices_per_date.xlsx', index_col=0)

np.isnan(df_indices.reindex(['2015-01-02']).COEF_ANTERIOR[0])

df_indices.index.min(), df_indices.index.max()

df_articles_sem_merge.head()

df_articles_sem_merge.to_csv('labeled_articles_sem_merge.csv', sep='\t')

df_articles_etiquetados.head()

df_articles_sem_merge.date.min(), df_articles_sem_merge.date.max()

# df_merged = merge_df(df_articles, df_indices)

df_sem_merge_dos_artigos = merge_df(df_articles_sem_merge, df_indices)

df_sem_merge_dos_artigos[df_sem_merge_dos_artigos.coef.isna()]

df_sem_merge_dos_artigos[df_sem_merge_dos_artigos.coef.isna()].shape, df_sem_merge_dos_artigos.shape

len(df_sem_merge_dos_artigos[df_sem_merge_dos_artigos.indice == 1]), len(df_sem_merge_dos_artigos[df_sem_merge_dos_artigos.indice == -1]), len(df_sem_merge_dos_artigos[df_sem_merge_dos_artigos.indice == 0])

df_sem_merge_dos_artigos[df_sem_merge_dos_artigos.title.str.contains('etrobras')].shape, df_sem_merge_dos_artigos.shape

df_sem_merge_dos_artigos.to_csv('labeled_articles_sem_merge_data.csv', sep='\t', index=True)
