==> step6_modelos_predicao_ml.ipynb <==
import pandas as pd
import numpy as np

import nltk
from nltk.corpus import stopwords
from nltk.probability import FreqDist

from wordcloud import WordCloud, STOPWORDS, ImageColorGenerator

import matplotlib as mpl
import matplotlib.pyplot as plt

from sklearn.naive_bayes import GaussianNB, MultinomialNB, ComplementNB, BernoulliNB, CategoricalNB
from sklearn.linear_model import SGDClassifier
from sklearn import tree
from sklearn.feature_extraction.text import TfidfVectorizer, CountVectorizer
from sklearn.ensemble import RandomForestClassifier, AdaBoostClassifier, GradientBoostingClassifier
from sklearn.neural_network import MLPClassifier
from sklearn.model_selection import train_test_split
from sklearn.model_selection import cross_val_score

from sklearn import tree
from scipy.spatial import distance
import time
import funcoes_uteis

from gensim.models import word2vec

vect_bag = CountVectorizer()

df_articles = pd.read_csv('stemmed_words/labeled_and_stemmed_articles2.csv', sep='\t')


df_articles.date.min(), df_articles.date.max()

df_articles[df_articles.link.str.contains('mercado')].shape, df_articles.shape

df_articles[df_articles.indice == 0].shape

df_train = pd.concat([df_articles[df_articles.indice == 0].sample(19), df_articles[df_articles.indice == 1].sample(50), df_articles[df_articles.indice == -1].sample(50)], ignore_index=True)
# df_train = df_articles[df_articles.link.str.contains('mercado')]


tokens_stemmer = df_train.stemm

vect_bag = CountVectorizer()
docs_bag = vect_bag.fit_transform([' '.join(tokens_stemmer)])

frase_count_vec = []
frase_count_vec.append(docs_bag.todense()[0])

vect_bag.get_feature_names_out()



vec_contents = vect_bag.fit_transform(df_train.stemm).toarray()



X_train, X_test, y_train, y_test = train_test_split(vec_contents, df_train.indice)

vect_bag.get_feature_names_out()

gnb = GaussianNB(var_smoothing=1e-110)
gnb.fit(X_train, y_train)

cross_val_score(gnb, X_test, y_test, cv=3)

pd.Series(y_train).unique()

gnb = MultinomialNB(alpha=0.1)
gnb.fit(X_train, y_train)

cross_val_score(gnb, X_test, y_test, cv=3)

clf = tree.DecisionTreeClassifier()
clf = clf.fit(X_train, y_train)

cross_val_score(clf, X_test, y_test, cv=5, )

clf = RandomForestClassifier(criterion ='gini')
clf = clf.fit(X_train, y_train)

cross_val_score(clf, X_test, y_test, cv=3)

clf = AdaBoostClassifier()
clf = clf.fit(X_train, y_train)

cross_val_score(clf, X_test, y_test, cv=3)

clf = GradientBoostingClassifier(n_estimators=100, learning_rate=1.0, max_depth=1, random_state=0)
clf = clf.fit(X_train, y_train)

cross_val_score(clf, X_test, y_test, cv=3)

clf = SGDClassifier()
clf = clf.fit(X_train, y_train)

cross_val_score(clf, X_test, y_test, cv=3)


clf = SGDClassifier(loss='modified_huber', )
clf = clf.fit(X_train, y_train)

cross_val_score(clf, X_test, y_test, cv=3)


clf = SGDClassifier(loss='perceptron')
clf = clf.fit(X_train, y_train)

cross_val_score(clf, X_test, y_test, cv=5)
